Added an svgbtnwidget class and a back button to go with it!
I moved the player svgs into a player folder and the new svg to a UI folder under the original assets folder.
This impacted the player view (as it needed to use the new path) and the tasksview (to add the new button).
Next I will add a custom graphics view to have better scrolling and such