Fixed bugs with window sizing; the window now correctly uses floating point values, and the graphicsviewcanvas handles transforming correctly
Modified the window class to use 100.0f instead of 100 and QPointF instead of QPoint for the floating point values and added an offsetpos func to the graphicsviewcanvas to handle scrolling, as otherwise it has a bad time.
This had no other impacts apart from fixing these bugs.
Next I'll make the tasks widgets expand on hover!