Improved the taskwidgets' todos; making them sync with the parent and other small changes!
Made the taskwidgets' todos pass through events when inactive, have the right cursor and now update the parent too!
Also, removed the click label to change the checkbox state; as that made editing the label harder.
Additionally improved the 'when finished editing' to be a focusOutEvent function.
This affects the taskswidget files.
Next I will make users able to delete tasks and subtasks.