Updated every file to have improved, standardised comments and docstring
Also realised that my variable naming conventions were mostly fine in the first place.
But the documentation was all over the place!
Now every function (but the initialisers) have a docstring, most functions have comments explaining everything and everything is nice and neat.
This impacted every file but was crucial as it was the next step through in development.
Next I will move the todo graphic objects into their own file, the task widget file is getting bloated with partially related classes.