Some bugfixes
Task overlay's graphics view now updates along with task widget
added a minimum maximum zoom (so as to with little tasks not be almost touching them)
Fixed a crash with the section bg widget when it's empty
Fixed a bug to do with expanding when clicking on a task widget
Other minor things

These affected the graphics view canvas, section bg widget, task widget and task overlay files.
These are very important as they affect the user's usage of these new and existing features.
*Next* I'll fix the todo label space issue.